# Phase 2/specs/002-responsive-frontend/contracts/frontend-openapi.yaml
openapi: 3.0.0
info:
  title: Todo Frontend API Integration
  version: 1.0.0
  description: API contracts for frontend integration with Todo backend
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
  /api/v1/tasks:
    get:
      summary: Get all tasks for authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of tasks for the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - insufficient permissions
    post:
      summary: Create a new task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired token

  /api/v1/tasks/{task_id}:
    get:
      summary: Get a specific task by ID
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - user does not own this task
        '404':
          description: Task not found
    put:
      summary: Update an existing task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - user does not own this task
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - user does not own this task
        '404':
          description: Task not found

  /api/auth/signup:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/signin:
    post:
      summary: Authenticate existing user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      summary: Logout user and invalidate session
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Unauthorized - invalid or expired token

  /api/users/me:
    get:
      summary: Get current user information
      tags:
        - User
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized - invalid or expired token
        '403':
          description: Forbidden - insufficient permissions

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          minLength: 8
          description: User's password
          example: securePassword123

    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        password:
          type: string
          description: User's password
          example: securePassword123

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          enum: [bearer]
          description: Token type
          example: bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 900
        user_id:
          type: integer
          description: Unique identifier of the authenticated user
          example: 123
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier of the user
          example: 123
        email:
          type: string
          format: email
          description: User's email address
          example: user@example.com
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2023-10-20T10:00:00Z"

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Task description
          example: "Get milk, bread, and eggs"
        priority:
          type: integer
          description: Task priority level (1-5)
          minimum: 1
          maximum: 5
          default: 3
          example: 2
        due_date:
          type: string
          format: date-time
          description: Task due date in ISO format
          example: "2023-12-31T23:59:59Z"
        is_completed:
          type: boolean
          description: Task completion status
          default: false

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title
          example: "Updated task title"
        description:
          type: string
          description: Task description
          example: "Updated task description"
        is_completed:
          type: boolean
          description: Task completion status
          example: true
        priority:
          type: integer
          description: Task priority level (1-5)
          minimum: 1
          maximum: 5
          example: 4
        due_date:
          type: string
          format: date-time
          description: Task due date in ISO format
          example: "2023-12-31T23:59:59Z"

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 123
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Task description
          example: "Get milk, bread, and eggs"
        is_completed:
          type: boolean
          description: Task completion status
          example: false
        priority:
          type: integer
          description: Task priority level (1-5)
          example: 2
        due_date:
          type: string
          format: date-time
          description: Task due date in ISO format
          example: "2023-12-31T23:59:59Z"
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2023-10-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2023-10-20T10:00:00Z"
        user_id:
          type: integer
          description: ID of the user who owns the task
          example: 456
        is_deleted:
          type: boolean
          description: Whether the task has been soft deleted
          example: false

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        total_count:
          type: integer
          description: Total number of tasks for the user
          example: 5
        completed_count:
          type: integer
          description: Number of completed tasks
          example: 2

    DeleteResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: "Task deleted successfully"
        deleted_id:
          type: integer
          description: ID of the deleted task
          example: 123

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid input parameters"